Add pre-dump functionality
This adds an option to pre-dump a container before doing the final dump.
For big applications this could reduce the amount of time the app is not
able to serve requests while checkpointing. A benchmark has also been
added to see the impact pre-dump has. It seems to have a slight impact
on restore performance, so it should probably only be used for
applications where dumping is really slow without a pre-dump. Because of
that and because arm lacks pre-dump support it's opt-in.

https://criu.org/Memory_changes_tracking