Track last socket activity on process for scheduling scaledown
This introduces an ebpf-based socket tracker to help zeropod making
the decision on when to checkpoint the container. Only if there was
no activity for however long the ScaleDownDuration is set, it will
checkpoint the container. If there has been some activity, it will
be delayed. This reduces the chances of checkpointing while someone
is actively using the service.