Add metrics
This adds metrics to zeropod by exposing a unix socket on each shim that
returns metrics of the current process. To collect these metrics the
manager container is introduced as part of the zeropod node daemonset (which
also includes the installer). The manager iterates all metrics sockets,
merges the metrics and exposes them on a normal http `/metrics` path.