feat: consider all containers by default
with this change, users don't need to specify which container is getting
scaled down, if nothing is specified, every container will be
considered. The container names can still be specified to select
specific container(s). With this, zero configuration is required via
annotations so zeropod will work by just setting the runtimeClass.