fix: improve auto ports detection
as checking just /proc/net/tcp will give us all listening sockets
in the current net ns, we need to do some additional matching to find
out which container in our pod is listening on which port in case we
have multiple containers. This is done by matching the socket fd with
ones listed in /proc/<pid>/fdinfo. So far this has worked pretty
reliably. In case this fails though, it's also possible for the user
to specify the container/port mapping manually using an annotation.
While at it, the annotations have been refactored for consistency.