fix: activator redirect loops
To avoid redirect loops, the activator now uses a known port for the
connection to the backend so we can disable redirects for these
packets.

Additionally this splits the bpf program into two separate programs for
ingress and egress, mainly to make things easier to understand but it
also makes the egress path shorter.