feat: improve installer/uninstaller
the installer is now being run in an initContainer of the manager. It
wasn't before as the installer existed before we had the manager pod.

The uninstaller is now making use of the same daemonset as we need to
run it on all nodes, just like the installer. It simply changes some
details of the installer manifest to make sure it runs with the
uninstall flag.