feat: move redirector loading to node daemon
Loading the redirector eBPF program directly in the shim is not the best
idea since program loading allocates quite a bit of memory. The memory
can be freed again by calling btf.FlushKernelSpec() but since the Go
runtime will not reduce the heap after that, we end up with shims that
GC less frequently, making them more memory hungry overall.

As a side-benefit, this makes it possible to change out the redirector
programs without having to restart workloads. Of course special care
needs to be taken that the map structure is still compatible as the
shims still need to access the maps.