fix: handling process exits after restore
When deleting the container at a point in time where the process is
restored our shim would not get the exit signal. To fix this we need to
setup the process handlers on restore as if it was a new container. We
do this by hooking into the lifecycle with the preRestore and
postRestore functions. This allows us to call the relevant task service
functions to setup the exit handlers after every restore.