feat(experimental): add grouping of multiple pods on one shim
Multiple pods can be grouped into one shim by setting the annotation
`io.containerd.runc.v2.group: "zeropod"`. The value defines the group
id for each of which one shim process is started. This commit ensures
that annotation is passed to the shim and also fixes the metrics to
properly cleanup when grouping and pods are removed.

This is currently marked as experimental since not much testing has been
done and new issues might surface when using grouping.