feat: implement in-place pod request scaling
If enabled, we make use of the (still alpha) feature gate
InPlacePodVerticalScaling to change the pod requests to a minimum on
scaling down and increasing them to the original value on restore. This
has been marked as experimental and also needs to be enabled by setting
a flag on the manager and of course the feature flag needs to be
enabled.