ci: use branch to prepare for release
This changes the release process a bit. Instead of pushing a tag and
then committing image version updates with a commit after the tag, the
image version changes are done with the same commit as the tag. This
makes it possible to checkout a tag and modify the deployment manifests
while having the correct version set on the images. The process will be
as follows:

* git checkout -b release/v0.4.0 && git push
* action updates and commits the image versions
* action creates a PR with the version changes
* after merge, commit can be tagged and pushed