fix: race condition with process exits
This gets rid of an older workaround where we set the container to scaled
down before checkpointing to avoid the exit caused by the checkpoint to
be processed. Instead we now keep track of PIDs which we expect to be
checkpointed and then look up the PID when processing an exit to tell if
the exit was related to a checkpoint instead of any other reason.