feat: add support for resizing on Kubernetes 1.32
The resizing of running pods has changed slighly with 1.32. There's a
new feature gate that needs to be enabled to get the containerStatuses
field. Also instead of just updating the pod spec, the resource update
now needs to be done using the pods/resize subresource. To be compatible
with 1.27-1.31, it will fall back to using a normal update call if the
resize fails with a not found.