fix: improve attacher reliability in kind
This improves kind (and with it e2e test) reliability for the redirector
attacher. As multiple kind nodes share the same kernel and bpf
environment, the attacher would sometimes attach on the wrong node,
leading to removals of the redirector which broke things. To avoid this,
the redirector will now check if the net ns actually exists within their
env and only attach it then.