feat: use containerd v2 for building shim
To be able to support new containerd features such as user namespaces,
the shim needs to use the containerd v2 manager. There are still various
components that depend on containerd v1 that we can't get rid of right
now as they moved the packages to internal. The log/io restoring is the
biggest problem here but this should be fine for the time being.

Also to be still able to use the shim with containerd v1, the shim just
registers as a task v2 service instead of the new v3. As v3 has not changed
anything this should also be fine for now. Once containerd v2 is widely
adopted, we can switch to task v3 but right now there don't seem to be
any advantages at all.