feat: move prometheus metrics to manager
To keep the shim binary and memory usage small, the prometheus package
is no longer imported in the shim. The metrics are simply stored in the
container struct and fetched by the manager via TTRPC. Only there they
are converted to prometheus metrics. This also simplifies the logic as
we no longer need to merge metrics from all shims, instead we just
record them on collect from all the shims.