fix: improve opt path configuration
If containerd configures another opt path, we don't change it to
/opt/zeropod as we don't want to break anything. But this instead
meant that our runtime binaries weren't installed in the correct
location. Another problem is the new criu usage in the manager where
we simply mounted /opt/zeropod from the host, which is not guaranteed.
To improve this, the installer will now read the opt path from the
containerd configuration first and installs the binaries there if
the opt path is preconfigured. For the manager, we simply put the criu
binary/libs into the container image, removing the need to mount a
hostpath, which is way cleaner anyways.