feat: implement migrating upper container layer
this is a very basic implementation of migrating the containers
ephemeral storage during migration. It only supports the overlay
snapshotter. It simply finds the upper layer by looking at
/proc/*/mounts and then renames all paths in the upper layer into
the snapshot directory before evac. On restore we do the reverse.
This works okay for a first iteration but there may be many
edge-cases and potential performance improvements. Especially the
image pulling over TTRPC could be improved.