feat: add restore ready field to migration
This adds a new field to the migration to indicate a migration is ready
for being restored. This is so the claiming and start of the checkpoint
can be separated. This allows us to offload the migration claiming to
the pod controller so we can quickly claim a migration while things like
pulling images are still in progress on the new pod. The evacuating pod
then knows there is a potential migration target and we can have a
longer timeout to wait until the image pulls are done.