feat: claim migration in pod controller
We now try to claim a migration as quickly as possible after the new pod
has been scheduled. The migration can still be claimed by the node
service if the controller failed to do so for whatever reason but in the
most cases the controller will be earlier so a migration can be claimed
before things like image pulling has been completed.