feat: merge socket tracker into redirector
This merges the two bpf programs into one. This has several advantages:

* we don't need to match container traffic via PIDs anymore as the
  activator already runs once per container. So we can just track
  activity in there.
* we no longer need to have a special PID resolver for running in kind.
* should be more reliable in general as long-running TCP connections
  will be tracked more accurately.

One disadvantage is that we now do more processing in the hot-path of
ingress traffic as we track activity on every packet, not just on every
TCP accept. Looking at bpftop the ingress path takes ~1700ns and egress
(which is unchanged) ~700ns. But the ingress path was already more
expensive before this change as it needs to do the disabled check. There
might still be room for optimization here.